+++ src/shunting/models/Event.java
@@ -1,15 +1,15 @@
 package shunting.models;
 
 public abstract class Event implements Comparable<Event> {
-	private double time;
+	private int time;
 	private Composition comp;
 	
-	public Event(double time, Composition comp) {
+	public Event(int time, Composition comp) {
 		this.time = time;
 		this.comp = comp;
 	}
 	
-	public double getTime() {
+	public int getTime() {
 		return time;
 	}
